requirements.txt
pydantic>=2.5
pytest>=8.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=1.0.0
langchain-core>=1.0.0
langchain-openai>=0.1.0
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langsmith>=0.1.0
supabase>=2.24.0
mangum>=0.17.0
boto3>=1.34.0
